stack: use byte array for channel masks
The RCP API has been maintained as it was. In fact using
spinel_push_fixed_u8_array results in the same byte order as using
spinel_push_fixed_u32_array.

A 32bit unsigned integer b31..b0 is written in little endian in spinel:

b7..b0 b15..b8 b23..b16 b31..b24

Which is the same as 4 8bit unsigned integers in order.

WI_SUN-1374

+++ common/parsers.h
@@ -15,7 +15,7 @@
 #include <stdint.h>
 #include <sys/socket.h>
 
-int parse_bitmask(uint32_t *out, int size, const char *str);
+int parse_bitmask(uint8_t *out, int size, const char *str);
 int parse_escape_sequences(char *out, const char *in, size_t max_len);
 int parse_byte_array(uint8_t *out, int size, const char *str);
 void parse_netaddr(struct sockaddr_storage *out, const char *str);
