+++ stack/source/6lowpan/ws/ws_common.h
@@ -140,9 +140,9 @@ typedef struct ws_info {
 } ws_info_t;
 
 
-int8_t ws_common_generate_channel_list(const struct net_if *cur, uint32_t *channel_mask, uint16_t number_of_channels, uint8_t regulatory_domain, uint8_t operating_class, uint8_t channel_plan_id);
+int8_t ws_common_generate_channel_list(const struct net_if *cur, uint8_t *channel_mask, uint16_t number_of_channels, uint8_t regulatory_domain, uint8_t operating_class, uint8_t channel_plan_id);
 
-uint16_t ws_common_active_channel_count(uint32_t *channel_mask, uint16_t number_of_channels);
+uint16_t ws_common_active_channel_count(uint8_t *channel_mask, uint16_t number_of_channels);
 
 int8_t ws_common_regulatory_domain_config(struct net_if *cur, ws_hopping_schedule_t *hopping_schedule);
